pyproject.toml
[project]
name = "2025-koudo-b"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "m-tsuru", email = "147146974+m-tsuru@users.noreply.github.com" }
]
dependencies = [
    "beautifulsoup4>=4.13.4",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "types-requests>=2.32.0.20250515",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
